@@ -1,20 +1,20 @@
 package qupath.ext.biop.cellpose;
 
 import javafx.beans.property.StringProperty;
+import org.controlsfx.control.PropertySheet;
 import org.controlsfx.control.action.Action;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import qupath.fx.prefs.controlsfx.PropertyItemBuilder;
 import qupath.lib.common.Version;
 import qupath.lib.gui.QuPathGUI;
 import qupath.lib.gui.extensions.GitHubProject;
 import qupath.lib.gui.extensions.QuPathExtension;
-import qupath.lib.gui.panes.PreferencePane;
 import qupath.lib.gui.prefs.PathPrefs;
 import qupath.lib.gui.tools.MenuTools;
 
 import java.io.InputStream;
-import java.util.Map;
-
+import java.util.LinkedHashMap;
 
 /**
  * Install Cellpose as an extension.
@@ -29,10 +29,10 @@ public class CellposeExtension implements QuPathExtension, GitHubProject {
 
     private boolean isInstalled = false;
 
-    private static final Map<String, String> SCRIPTS = Map.of(
-            "Cellpose training script template", "scripts/Cellpose_training_template.groovy",
-            "Cellpose detection script template", "scripts/Cellpose_detection_template.groovy"
-    );
+    private static final LinkedHashMap<String, String> SCRIPTS = new LinkedHashMap<>() {{
+        put("Cellpose training script template", "scripts/Cellpose_training_template.groovy");
+        put("Cellpose detection script template", "scripts/Cellpose_detection_template.groovy");
+    }};
 
     @Override
     public GitHubRepo getRepository() {
@@ -41,7 +41,7 @@ public class CellposeExtension implements QuPathExtension, GitHubProject {
 
     @Override
     public void installExtension(QuPathGUI qupath) {
-        if(isInstalled)
+        if (isInstalled)
             return;
 
         SCRIPTS.entrySet().forEach(entry -> {
@@ -61,9 +61,11 @@ public class CellposeExtension implements QuPathExtension, GitHubProject {
         // Get a copy of the cellpose options
         CellposeSetup options = CellposeSetup.getInstance();
 
+
         // Create the options we need
         StringProperty cellposePath = PathPrefs.createPersistentPreference("cellposePythonPath", "");
         StringProperty omniposePath = PathPrefs.createPersistentPreference("omniposePythonPath", "");
+
 
         //Set options to current values
         options.setCellposePytonPath(cellposePath.get());
@@ -73,13 +75,23 @@ public class CellposeExtension implements QuPathExtension, GitHubProject {
         cellposePath.addListener((v, o, n) -> options.setCellposePytonPath(n));
         omniposePath.addListener((v, o, n) -> options.setOmniposePytonPath(n));
 
-        // Add Permanent Preferences and Populate Preferences
-        PreferencePane prefs = QuPathGUI.getInstance().getPreferencePane();
+        PropertySheet.Item cellposePathItem = new PropertyItemBuilder<>(cellposePath, String.class)
+                .propertyType(PropertyItemBuilder.PropertyType.GENERAL)
+                .name("Cellpose 'python.exe' location")
+                .category("Cellpose/Omnipose")
+                .description("Enter the full path to your cellpose environment, including 'python.exe'")
+                .build();
 
-        prefs.addPropertyPreference(cellposePath, String.class, "Cellpose 'python.exe' location", "Cellpose/Omnipose",
-                "Enter the full path to your cellpose environment, including 'python.exe'");
-        prefs.addPropertyPreference(omniposePath, String.class, "Omnipose 'python.exe location (Optional)", "Cellpose/Omnipose",
-                "Enter the full path to your omnipose environment, including 'python.exe'");
+        PropertySheet.Item omniposePathItem = new PropertyItemBuilder<>(omniposePath, String.class)
+                .propertyType(PropertyItemBuilder.PropertyType.GENERAL)
+                .name("Omnipose 'python.exe' location")
+                .category("Cellpose/Omnipose")
+                .description("Enter the full path to your omnipose environment, including 'python.exe'")
+                .build();
+
+        // Add Permanent Preferences and Populate Preferences
+        QuPathGUI.getInstance().getPreferencePane().getPropertySheet().getItems().addAll(cellposePathItem, omniposePathItem);
+
     }
 
     @Override
@@ -103,6 +115,6 @@ public class CellposeExtension implements QuPathExtension, GitHubProject {
             logger.error("No script editor is available!");
             return;
         }
-        qupath.getScriptEditor().showScript("StarDist detection", script);
+        qupath.getScriptEditor().showScript("Cellpose detection", script);
     }
 }
